python-backend/requirements.txt
flask==3.0.0
flask-cors==6.0.1
mysql-connector-python==9.4.0
